Исправление замечаний по ДЗ
