+++ Navigation/Navigation/Post.swift
@@ -17,6 +17,7 @@ struct Post {
     static func makePost() -> [Post] {
     
         var post = [Post]()
+        
         post.append(Post(author: "Киновселенная Марвел", image: UIImage(named: "TomHolland")!, description: "Том Холланд вернётся к роли Человека-паука в MCU. Информация о том, что Том Холланд не закончил с партией Питера Паркера после отгремевшего триквела «Нет пути домой», уже появлялась, но никакой конкретики вовлечённые в процесс принятия решений лица не сообщали. Нынче же кое-чем может поделиться сам Кевин Файги — глава Marvel Studios.", likes: Int.random(in: 300...500), views: Int.random(in: 700...1999)))
         post.append(Post(author: "Star Wars", image: UIImage(named: "Mandalorian")!, description: "Гриф Карга уговаривает Дина Джарина в первом отрывке из третьего сезона «Мандалорца». В промо-кампанию третьего сезона «Мандалорца» вступают непосредственные отрывки из свежих эпизодов. В сети появился первый фрагмент из продолжения, где Гриф Карга (Карл Уэзерс) пытается уговорить Дина Джарина (Педро Паскаль) остаться и принять новый заказ, но у главного героя другие планы. Мандалорец продолжит свой непростой путь, который рано или поздно приведёт его на родину. В путешествиях Дина Джарина будет сопровождать не только Грогу. Трейлеры и промо-кадры прозрачно намекали, что вместе с ним какое-то время попутешествует Бо-Катан Крайз (Кэти Сакхофф).", likes: Int.random(in: 300...500), views: Int.random(in: 700...1999)))
         post.append(Post(author: "Star Wars", image: UIImage(named: "Andor")!, description: "Диего Луна и Стеллан Скарсгард засветились на съёмках второго сезона «Андора». Один из лучших сериалов прошлого года вернулся на производственные рельсы. В столице Великобритании не так давно стартовало активное производство второго сезона «Андора», и в сети уже появились фотографии со съёмочной площадки новых эпизодов. На снимках засветились ключевые персонажи истории — Кассиан Андор (Диего Луна) и Лютен Раэль (Стеллан Скарсгард).", likes: Int.random(in: 300...500), views: Int.random(in: 700...1999)))
