+++ Navigation/Navigation/Profile/ProfileViewController.swift
@@ -28,7 +28,7 @@ final class ProfileViewController: UIViewController {
     
     private func layout() {
         view.addSubview(tableView)
-        tableView.tableHeaderView = ProfileTableHeaderView(frame: CGRect(x: 0, y: 0, width: tableView.frame.width, height: 250))
+        tableView.tableHeaderView = ProfileTableHeaderView(frame: CGRect(x: 0, y: 0, width: tableView.frame.width, height: 220))
         tableView.backgroundColor = .systemGray4
         
         NSLayoutConstraint.activate([
